+++ beam/beam-core/src/main/java/org/esa/beam/util/geotiff/GeoCoding2GeoTIFFMetadata.java
@@ -131,8 +131,8 @@ public class GeoCoding2GeoTIFFMetadata {
 
         GeoPos geoPos1 = geoCoding.getGeoPos(new PixelPos(0.5f, 0.5f), null);
         GeoPos geoPos2 = geoCoding.getGeoPos(new PixelPos(1.5f, 1.5f), null);
-        final double scaleX = Math.abs(geoPos2.lon - geoPos1.lon);
-        final double scaleY = Math.abs(geoPos2.lat - geoPos1.lat);
+        final float scaleX = Math.abs(geoPos2.lon - geoPos1.lon);
+        final float scaleY = Math.abs(geoPos2.lat - geoPos1.lat);
         metadata.setModelPixelScale(scaleX, scaleY);
         return metadata;
     }
