+++ beam/beam-ui/src/main/java/org/esa/beam/framework/ui/WorldMapPane.java
@@ -461,8 +461,8 @@ public class WorldMapPane extends JPanel {
             g2d.setColor(Color.black);
 
             g2d.drawString(text,
-                           (float)textCenter.x - fontMetrics.stringWidth(text) / 2.0f,
-                    (float)textCenter.y + fontMetrics.getAscent() / 2.0f);
+                           textCenter.x - fontMetrics.stringWidth(text) / 2.0f,
+                           textCenter.y + fontMetrics.getAscent() / 2.0f);
             g2d.setColor(color);
         }
