+++ beam/beam-visat-rcp/src/main/java/org/esa/beam/visat/actions/CreateXYDisplacementBandsAction.java
@@ -89,8 +89,8 @@ public class CreateXYDisplacementBandsAction extends ExecCommand {
         bandY.setUnit("pixels");
         bandY.setDescription("Geo-coding Y displacement");
 
-        final double[] dataX = new double[width * height];
-        final double[] dataY = new double[width * height];
+        final float[] dataX = new float[width * height];
+        final float[] dataY = new float[width * height];
 
         bandX.setRasterData(ProductData.createInstance(dataX));
         bandY.setRasterData(ProductData.createInstance(dataY));
