+++ beam/beam-visat-rcp/src/main/java/org/esa/beam/visat/toolviews/placemark/PlacemarkDialog.java
@@ -188,7 +188,7 @@ public class PlacemarkDialog extends ModalDialog {
         return (Float) bindingContext.getPropertySet().getValue(PROPERTY_NAME_PIXEL_X);
     }
 
-    public void setPixelX(double pixelX) {
+    public void setPixelX(float pixelX) {
         bindingContext.getPropertySet().setValue(PROPERTY_NAME_PIXEL_X, pixelX);
     }
 
@@ -196,7 +196,7 @@ public class PlacemarkDialog extends ModalDialog {
         return (Float) bindingContext.getPropertySet().getValue(PROPERTY_NAME_PIXEL_Y);
     }
 
-    public void setPixelY(double pixelY) {
+    public void setPixelY(float pixelY) {
         bindingContext.getPropertySet().setValue(PROPERTY_NAME_PIXEL_Y, pixelY);
     }
 
@@ -204,7 +204,7 @@ public class PlacemarkDialog extends ModalDialog {
         return (Float) bindingContext.getPropertySet().getValue(PROPERTY_NAME_LAT);
     }
 
-    public void setLat(double lat) {
+    public void setLat(float lat) {
         bindingContext.getPropertySet().setValue(PROPERTY_NAME_LAT, lat);
     }
 
@@ -212,7 +212,7 @@ public class PlacemarkDialog extends ModalDialog {
         return (Float) bindingContext.getPropertySet().getValue(PROPERTY_NAME_LON);
     }
 
-    public void setLon(double lon) {
+    public void setLon(float lon) {
         bindingContext.getPropertySet().setValue(PROPERTY_NAME_LON, lon);
     }
