+++ beam/beam-visat-rcp/src/main/java/org/esa/beam/visat/toolviews/placemark/pin/PinTableModel.java
@@ -53,8 +53,8 @@ public class PinTableModel extends AbstractPlacemarkTableModel {
         Assert.notNull(getProduct());
         final Placemark placemark = getPlacemarkDescriptor().getPlacemarkGroup(getProduct()).get(rowIndex);
 
-        double x = Float.NaN;
-        double y = Float.NaN;
+        float x = Float.NaN;
+        float y = Float.NaN;
 
         final PixelPos pixelPos = placemark.getPixelPos();
         if (pixelPos != null) {
@@ -62,8 +62,8 @@ public class PinTableModel extends AbstractPlacemarkTableModel {
             y = pixelPos.y;
         }
 
-        double lon = Float.NaN;
-        double lat = Float.NaN;
+        float lon = Float.NaN;
+        float lat = Float.NaN;
 
         final GeoPos geoPos = placemark.getGeoPos();
         if (geoPos != null) {
